+++ nix/nixos/de.6nw/pve.boreth/nix-builder/srv-autoUpdate/service.sh
@@ -46,7 +46,7 @@ git switch --force-create "$repoWorkingBranch"
 gcrootsDir="$CFG_gcrootsDir"
 gcrootsSuccess="$gcrootsDir/success"
 gcrootsWIP="$gcrootsDir/working"
-if [[ -d "$gcrootsWIP" ]]; then
+if [[ -e "$gcrootsWIP" ]]; then
     rm --recursive "$gcrootsWIP"
 fi
 mkdir --parent "$gcrootsWIP"
@@ -57,7 +57,9 @@ export CI_GCROOT="$gcrootsWIP"
 
 if ./update.sh && ./tests.sh; then
     # when successfully finished
-    rm --recursive "$gcrootsSuccess"
+    if [[ -e "$gcrootsSuccess" ]]; then
+        rm --recursive "$gcrootsSuccess"
+    fi
     mv "$gcrootsWIP" "$gcrootsSuccess"
     git branch --force "$repoDestBranch"
 fi
